refactor(paginate): move paginate into its own file
index.ts becomes pure re-exports, matching the one-unit-per-file layout.
